Added more token implementations and updated Token ABC Classes
