+++ paraccompiler/core/__init__.py
@@ -2,6 +2,9 @@
 """ Core of the Para-C Compiler (lexing, parsing, compiling, processing) """
 from . import optimiser
 from . import parser
+from . import abc
+from . import err_handler
+from . import logicstream
 from .compilation_ctx import *
 from .compiler import *
 from .. import *
