+++ paraccompiler/core/err_handler.py
@@ -9,7 +9,7 @@ from antlr4.error.Errors import (InputMismatchException,
                                  LexerNoViableAltException,
                                  NoViableAltException)
 
-from ..base_err_handler import BaseErrorListener
+from .abc.base_err_handler import BaseErrorListener
 
 __all__ = [
     'ParacErrorListener',
