+++ preprocessor/abc/__init__.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+""" ABC Classes used in the Pre-Processor """
+from . import base_token
+from .base_token import *
+
+__all__ = [
+    'base_token',
+    *base_token.__all__
+]
